Manager: Avoid deallocating streams that are currently running
Previously, the logic behind deallocating streams only took into
account if the player index had reached the end of the array, and
started deallocating from its first position onwards. This was a
problem because it is possible that some of those positions would
still be running when FreeJ2ME went to deallocate them.

Now it will also take into account whether the current position
is running or not, and will only forcefully deallocate a position
that's running if the entire array is allocated, and all the
players are running, to make space for a new one.