Fix LCD paint logic - use the repaint method
On mac, updates might not be performed when calling paint directly,
and the Graphics class shouldn't be retrieved directly.

Also on this TASEmulators fork, optimize FreeJ2ME's
lcd paint() method to only copy the 'Graphics g' argument into
a Graphics2D if screen rotation is enabled, which
should save us some time in the render path.