FreeJ2ME: Introduce support for audio stream dumping
Mostly for debugging purposes, but it's also neat to have the
ability to dump audio data from J2ME apps, especially ones that
are encrypted and can't be easily opened up.

Supported on both AWT and Libretro. As for Anbu/SDL... that one
is in need of a UI rewrite, as it's still too barebones to have
features be tacked onto it.

The dump file path is (related to the jar): FreeJ2MEDumps/Audio/appname/*