docs: Overhaul the webpage to reflect this fork.
FreeJ2ME-Plus should be treated as its own thing, and since the
original repo still doesn't have its own page as of yet, using
their icons and references could mislead some users into thinking
that this is the original since it is the most "complete".

This also comes with an all-new logo for FreeJ2ME-Plus as well,
which is also lighter than the previous one (partially because
a simplify pass was run over it, reducing the render accuracy
slightly, but not to a noticeable degree unless highly zoomed in).
The faster this webpage loads, the better.

As a plus, the webpage is in "dark mode" now.