Improve Custom Soundfont loading, also fix an AWTGUI bug with it
Previously AWTGUI was checking against the "soundfont" field being
"on" or "off" instead of "Custom" or "Default". This is now fixed.

Also, the actual soundfont loading procedure is now more cross-platform
and less error-prone, falling back to the default JVM font if no
sf2 fonts are found. Anbu/SDL2 also has received code to use this
feature as well... still in dire need of a UI rework though.