FreeJ2ME: Add a few more common-ish resolutions to the list
This commit also removes 96x96, as there is no phone running this
to my knowledge.

As for the added resolutions, they are:

101x64  (Siemens M50)
101x80  (in place of 104x80, which no phone seems to use either)
130x130 (Some Siemens C-Series)
128x142 (Some Sagem my-Series phones)
132x176 (Siemens CX65)
176x204 (Motorola V-Series)
220x176 (Samsung X820)
208x320 (Sony Ericsson P-Series)
240x432 (Sony Ericsson Aino)
240x480 (LG KF700)
640x480 (Nokia E6-00)

Of course, there are still some less conventional screens in the
wild, but many of them are either mostly covered by what we already
have, or are seen in feature phones that don't sport MIDP support
at all.