Libretro: Use a byte image type for the canvas image
This allows us to get the databuffer directly in byte[] form,
which means that manipulations can be done directly to it while
skipping costly calls to BufferedImage.getRGB().

In general, doesn't make much of a difference, although jBenchmark
reports 8% higher performance on average (9061 vs. 8376, i5-8250U)

At least it's closer to the standalone jar now, which scores an
average of about 9826 points in the same test platform. There's always
going to be an overhead due to how the libretro interface works though.