Anbu: Improve on the new SDL interface.
This should not only help it render normally with a workaround (entire
window was a black canvas on my platform before), but also fix
some other issues it had, like having input logic tied to render
logic, and not supporting window scaling + fullscreen (this one still
needs work).

It now has support for Dumping Midi Streams and Custom SoundFonts,
although the user can't set it at runtime due to the lack of a UI,
which hopefully won't be much of a problem if AWT's integration
ends up being easy enough.

Another "issue" for now is performance, with this interface performing
FAR worse than the Standalone AWT jar, and Libretro's interface, but
it's also something that can be worked on later.

Dependencies are also added to the repo with this one. They won't need
to be updated any time soon, and might even be downgraded to make
sure we support older versions of SDL2.