lcdui: Sprite: Refactor to improve code clarity
Some parts of this could be shortened up quite a bit, like methods
that had separate versions for X and Y coordinates, not to mention
my last commit on this one not really tidying up as well as it
could've. This one should deal with the bulk of it, and doesn't
introduce any breaking changes... at least i don't think so, given
the output should be the same.