lcdui: Rewrite major parts of Sprite, TiledLayer and LayerManager
Further improvements to TiledLayer stuff, i'm content with their
state now, as long as they remain as compatible as J2ME-Loader's
reference implementation... now for the commit message itself:

Rewrites were made in order to improve code clarity and readability
by turning extensive calculations into helper functions to better
delineate what is happening on the bigger functions inside Sprite
an TiledLayer, as well as general variable tightenings and
formatting changes. Thanks to this rewrite, TiledLayer is ever so
slightly faster in JBenchmark 2, but it's almost within margin
of error so it doesn't matter.

As far as functionality goes, none of the changes should break
compatibility and output in any way, although none of the
CollidesWith methods were tested, as i couldn't find a jar that
uses any of them... debug messages were left on all three so that we
can trace any weirdness in TiledLayer down to them, in case they're
the issue.