media: Refinements to the Media Cache
Sonic Jump (Nokia, 240x320) among many other MIDlets do deallocate
their streams to make room for other streams to save memory on
real hardware... this, however, is a problem when we're dealing
with a cache of any sort. We can't let those deallocate calls
actually go through or else the cache will, at best, be less useful,
or at worst, cause trouble as it'll think it still has an allocated
player in a certain cache position while the MIDlet has deallocated it.

So to deal with this, standard deallocate() calls were made into
glorified stop() calls, with only the Manager itself being able
to actually deallocate streams from the cache.

As a minor note: TonePlayer now has a debug message in its constructor
to help find cases where MIDlets are trying to use it.