WavImaAdpcmDecoder: Improve things a little bit further
We don't need to instance this Decoder every time, just make it
static and call directly whenever needed. Also, IMA ADPCM header
appears to be 60 bytes long, so account for that, and also refine
decoding a tiny bit by not accepting out-of-bound values for table
indices from the preamble just to clamp them arbitrarily to either
0 or 88. Use a simple approximation of what the step should be
based on the difference between the last decoded sample and the
one read from the preamble in those cases.

To be fair, this doesn't seem to be working out, shelving it
for another implementation might be a better call.