Libretro: Improve restart routines
It now loads any updated settings, whereas previously it just
ignored any updates and left them to a complete close and reopen.

Which kinda defeated the point of having the restart there as a
quicker way of applying settings that require it.

NOTE: Only tested on linux, although it should also work on Windows