PlatformPlayer: realize() again after deallocate(), conditionally
deallocate() can be called between the transition from UNREALIZED
to REALIZED, which is the only situation where the player can
return to UNREALIZED. That case was already covered. What was not
covered, was realizing the player again after deallocate() if
it was on the REALIZED state or higher.

And since close() has no need for a realize() call (which WILL be
done in the superclass unless UNREALIZED), close() will call the
actual player's deallocate() method directly before setting state
to CLOSED.

Fixes a bunch of gameloft games (mainly k800i versions with their
more robust audio system) that were already fixed with the big
PlatformPlayer rewrite, but regressed temporarily as the players'
state handling was tightened.