tweaks needed for libtas
Normally, input reading should not be tied to the render logic, because that would softlock jars that only send new render data after an input is registered (Ex: JBenchmark 2 and some other jars that use Form UI).

But for determinism in libTAS, inputs must be pulled synchronously with rendering, because libTAS processes inputs (and a lot of other things) during frame boundaries, which is when the program calls its render function.