Improve FPS Limit again, and make it frontend-independent
We're now using LockSupport.parkNanos to achieve a more consistent
fps lock that works in the nanosecond domain as opposed to
Thread.sleep()'s milliseconds. This also has the added bonus of
not needing to enter a try-catch block. Still needs some work because
locking to 60 doesn't result in a perfect 60fps counter in JBenchmark 2
(not much else that really allows one to test this), but it's still
an improvement.

On another note, formatting probably needs some work, as this
commit is changing stuff that i personally didn't even touch.