Massively improve logging.
For starters, now it's toggleable, and also allows offloading to a
file. Class and package names are prepended into the message for
better tracing , and log levels improve readability by giving
insight on which entries are of higher priority (assuming your
editor can color those entries based on their priority, it's
even clearer).

AdpcmDecoder got a slight rework on its header read, and libretro had
a small bug on the customMIDI option which was fixed here too.

SDL/Anbu is still behind of course.