Introduce FreeJ2ME's first Speed Hack! (No Alpha on Blank Images)
This option makes FreeJ2ME create all blank images without an alpha
channel, cutting out alpha processing on all of them, which of
course also includes the emulated "LCD" screen.

Despite its name and function, it appears to be relatively safe
despite providing a ridiculous speed up that often goes from 50%
to 150%(!) faster as far as rendering goes, with the slower cases
being software 3D rasterization that often doesn't rely much on Java
Image APIs in general.

Will be relegated to a hack because no matter how unlikely, there's
bound to be a jar that creates a blank image just to paint it with
transparency at some point... and it also deviates from J2ME spec.