Libretro: We don't need a separate surface to render
Yet another round of lowering FreeJ2ME's libretro overhead.

Turns out that instead of drawing to a separate 3BYTE_BGR surface,
it's faster to just get the current MobilePlatform's "LCD" and
poll data from it directly from either INT_RGB or INT_ARGB then
convert it to a byte array with WritableRaster before sending to
libretro's core. This had the effect of removing any need for a
graphics object in Libretro.java as well, which then made the
custom Runnable useless as well, so all of that was cleaned up.

No color conversions are done on the core itself now, it's back
to converting in the Java side of things again, in a very similar
way to how it was done originally. Except this time the overhead
is so low that it might as well be considered on par with FreeJ2ME's
standalone AWT jar, which is the fastest available.

Anbu also got part of its comment about DataBuffer removed, as it
doesn't really differ from Libretro anymore.