microedition.media: Update PlatformPlayer and Manager for Siemens
Although unlikely, there is a possibility that siemens jars may try
to create a Player from a DataSource, so introduce the method to
do so, logging for any uses. Also, AH-1 SeaBomber tries to load a
midi file with the "audio/x-mid" type, which wasn't being checked
against at all in both Manager and PlatformPlayer, so this has also
been taken care of.