PlatformImage: Use MUCH faster transforms.
This commit is centered entirely on fixing Sonic 2 Crash's terrible
performance. It makes very heavy use of pretty much all of the
available transforms to construct everything from the backgrounds,
to stage tiles, to the general sprites. Previously the amount of
calls to AffineTransform would slow this down considerably, often
making it drop to 1-2 FPS ingame. This is no longer the case, as
using direct byte manipulation on these transforms allows the game
to maintain its (still choppy) 12-ish FPS with a lot of CPU time to
spare.

As a small bonus, there is a simple image dumping function in there
for when i decide to implement it on a larger scale.