PlatformImage: Actually process alpha on createRGBImage
Calls to createRGBImage are casted into PlatformImage constructor's
PlatformImage(int[], int, int, boolean) here.

Also, why wasn't this crucial bit implemented when the rest of it was?

Anyway... fixes #5 Asphalt 2: Urban GT (finally!) and probably many
others. The noAlpha speedhack can also be applied here, since it
must assume that, on cases where processAlpha will be false, the
resulting image won't have any alpha channel data in it, so create
an image of type INT_RGB instead of INT_ARGB if the hack is enabled.

This adds some validation bits into getRGB as an extra.

Of course, there's bound to be jars that read from this non-alpha
data as if it was ARGB anyway.