PlatformPlayer: Workaround for pitch shifts and midi instruments
For some reason, often when a midi sequence stopped so that other
could play, and then starts again, it starts with notes in a
different pitch than expected, to the point of almost sounding
like a different instrument in certain cases (Sonic 2 Crash's
BGM into Invicibility jingle, then back to BGM for example).

This should fix such cases, without regressing much of anything.