MobilePlatform: When loading a jar, check for accompanying jad file
This is mostly a workaround for libretro's database. Since it will
match only against jar files, .jad files won't be added to the J2ME
playlist. So what we can do here is check if, when a jar is loaded,
there is an accompanying jad with the same name on the same directory.

If there is, we load it up, since often games benefit from having
its jad properties present, and others even refuse to get ingame
without them (certain versions of DOOM RPG). This still maintains
single jar loading, and direct jad loading as they were before.