PlatformPlayer: Minor improvements to state handling.
We shouldn't be able to retrieve the media's current time if
the player isn't at least prefetched (according to the docs, as
PREFETCHED is the only state where the exclusive player resources
should be fetched). Also, set the player state to PREFETCHED when
it reaches an END_OF_MEDIA event before notifying the playerListeners
about it.

Also, handle allocation, closing and deallocating a bit better
in regards to what the should actually do on this player
implementation, with deallocate() basically doing nothing, like
prefetch().

Fixes #16 (Sonic Jump Siemens version) and probably many others.