RecordStore: Make OpenRecordStore C a DEBUG call instead of WARNING
Sonic 2 Crash uses this, and it works, so no need to make it seem
like it's untested or something.

As a bonus, also add the record store's name on this one and its
siblings' debug calls.

Sidenote: RMS is another FreeJ2ME subsystem that needs some work
to better match the j2me docs, just like PlatformPlayer previously
had. But this is something for later.