EXPERIMENTAL: RecordStore: Rewrite a big chunk of the RMS subsystem
RecordStore operations have to be atomic. This was already covered
by using a Vector for records. What was not being covered were the
multiple instances where checks for an open record store had to
occur. It is checked now, with an implementation that's very similar
to J2ME-Loader. Beside this, some functions like getRecordId and
closeRecordStore weren't implemented... they are now.

Furthermore, the logic to delete and return the number of records
didn't make much sense (at least to me), so that too was rewritten.

There are probably a few more tidbits and changes here and there,
but the ones i've noted are the most important.

NOTE: Even if unlikely, there is still a possibility that this
might break some saves created on the previous implementation,
so beware. (Asphalt2, Sonic Jump, and Orcs & Elves had no issues)