PlatformPlayer: Implement almost all of MidiControl.
Shadow Shoot (again) seems to have weird audio playback on the
second half of the song that plays on Amethyst Castle, when the
church organ becomes the lead instrument. The hypothesis is that
it is issuing a program change of some sort in order to swap the
organ instrument during playback. So program changes, along with
the rest of MIDIControl sans getKeyName() were implemented to see
if that is indeed the case.