MelodyComposer: Try to match nokia's OTA/OTT tones correctness.
AH-1 SeaBomber has versions for both tone implementations, with
Nokia's newest implementation being assumed to be more correct,
as it is backed by some documentation. So tailor the sequence
BPM calculation and speed in siemens' MelodyComposer to be closer
to Nokia's on this game, hopefully improving other MelodyComposer
jars as well.

Note scales are now correct based on MIDI's expectations as well,
but they still sound very different from nokia... not much that
can be done here since Siemens is only supposed to go from
C0 (24) to A4 (81) directly, at least if Siemens' note definitions
and MIDI's note values are to match.