objectweb: Prune ASM by removing tons of unused code.
We only use a very small subset of ASM's functionality in FreeJ2ME,
so there is no need to have the whole library be present. By
removing everything that isn't required for our use case, ASM's
size when built is massively reduced and so is FreeJ2ME's resulting
jar. This also not only cuts the build time by a bit, but makes it
so there's a smaller surface area for issues to arise whenever
GraalVM gets considered.

More cuts can be done, but that would involve reworking quite a bit
of ASM's code, which i don't think is worth it, at least yet.