PlatformPlayer: More improvements to player state handling.
We should try realize() if a jar calls start() directly too. Also
actually close midi on deallocate to fix Castle of Magic's K800i
version. It'll be reopened on realize() and won't break Phoenix
Wright. Other than that, midi just had some small cleanups alongside
wavPlayer, which had its clip setup moved back to realize(). Death
Race only needs the player to be closed and then reopened, and
having wav be ready on realize() is better at the moment.

Copies of both decoded wav and normal wav are kept on the wav player
for cases like Castle of Magic, which deallocate the player and
expect to use it again with the same stream. Midi makes a bit more
sense now as well. It's state handling is a bit more robust overall.