Move libretro and (new) SDL platform checks into MobilePlatform
This also sees the introdution of a workaround for SDL to
force-draw at least 10 frames per second if the jar doesn't issue
any new redraw event. Useful in cases where the jar waits for
user input in order to draw anything new. Not generally a problem
in AWT and Libretro, but with SDL having input logic tied to the
rendering loop, not forcing redraws of any kind caused some jars
to softlock. JBenchmark2 is such a case.