PlatformPlayer: Make player creation more robust, and some cleanups
Now it should at least print about the format that the jar is
trying to load, and be more helpful even if the format isn't
supported yet, like SMAF and MFi for example.

Beyond that, remove some helper functions that i deem to be useless
now, while also making some small changes to formatting and variable
naming.