AudioClip: Basically ignore incoming media types.
This is too unreliable. Sonic 2 Crash is passing mmf streams as mp3
in some of its samsung versions, and now Spyro: The Eternal Night
has been found to be passing MIDI files as MMF.

So instead of using whatever the jar is passing as type, ignore
it entirely for all cases and parse the data's first bytes (often
the header, accounting for the possible audioOffset) to find out
what kind of data it actually is.