Use dataBuffers for other graphics workloads as well.
This does not break Munkiki's Castles, Crash Bandicoot, or any
other jar that was found to be buggy with the older, pre-fixed
get/setDataElements usage. So at least it's faster, and still
safe-ish.

On the other hand, creating a new transformed subimage using the
platformImage constructor now has a little more overhead since
the new subimage CANNOT share the same data array as its original
copy, otherwise the databuffer manipulation would make changes to
BOTH images. Hopefully it's not much of an overhead that it negates
the current optimizations, at least Castlevania: Dawn of Sorrow is
Still slightly faster than before, although it runs with a massive
performance margin even on an 800MHz i5-8250U.