PlatformPlayer + Manager: Make MIDI Synthesizer and Receiver static
This allows all sequencers (in short, every MIDI player requested
by any jar) to use the same, fixed references to synthesizer and
receiver. So instead of opening multiple synthesizers and getting
receivers whenever a MIDI is loaded, we cut all of it by just
getting the static synth's receiver every time.

Not only is it faster, but it also helps make midi volume changes
more functional in Java 8, with a small delay of 50ms being applied
to every receiver message to further help the sequencer follow through.