PlatformImage: Fix OOB access in MIRROR_ROT270, do subimage faster
The fun part of low-level optimizations, is that these kinds of
bugs get a lot more common, since Java's internal type and format
handling aren't here to save the day. Also, we can use a databuffer
to create a new subImage, so it is probably a bit faster now as
well.