Nokia: Sound: Don't return null in cases a command type is unknown
Some OTA/OTT tones are buggy and this is bound to happen. We can
just break out of the loop, as the tones might have been parsed,
even if partially. Such is the case of Ice Age's Nokia 3410 version.
Instead of crashing FreeJ2ME, it can now be played fine, although
sound can't be configured on the settings yet due to LCDUI being
barebones.