PlatformPlayer: Handle looping in a different way.
Neither MIDI nor WAV threw END_OF_MEDIA events once a loop was
completed. And because of that, neither was the STARTED event. So
now we don't actually set the loop count on the Sequencer and Clip
objects, instead the looping is managed by the metaEventListener
and LineListener added to the media once it's started, as it can
have access to notifying the player listener accordingly.

Sonic 1 Part One requires this in order to loop the invincibility
song and handle the game over transition correctly.

Hopefully, this won't break anything else.