WavImaAdpcmDecoder: More improvements to the decoder.
It appears that, whenever a malformed block is found, we should
actually stop decoding. But since this can still mean that the
decoder has decoded at least a partial frame of junk data, so
we still remove a bit of its final size in those cases just to be
sure.

This further reduces the noise in case a malformed block is found,
and doesn't seem to cut much into the decoded stream's sound, if
at all.