Manager + PlatformPlayer: Revamp dumps, try loading media locators
Opening media files from the jar when it sent a locator instead of
the inputstream and its type directly wasn't supported yet.

Also, media checks were simplified in both platformPlayer and
Manager. Although the inclusion of locators will end up with manager
dumping them with the incorrect extension depending on the name at
the moment. As a bonus, Nokia Sound also now exports the
original OTA tone sequence as well.