AudioClip: Assorted improvements to playback and state handling
The player has to be prefetched once it's created, otherwise making
volume changes on play() won't work. Also resume() should only
actually resume playing if the player is paused, AND it has not
reached the end of the media. play() will also behave like Nokia
Sound and always play from the beginning.