Canvas: Fix commandBar, make other classes less dependent on canvas
Now we don't even need a Canvas instance in MobilePlatform, as
all that's really needed to cast key mappings is a call to
Mobile.getGameAction(). This actually frees up Canvas itself to
receive improvements in how to handle its command Bar, which is
now rendered and interacted with correctly in far more cases.