Implement backlight flash for Display, Nokia, Samsung and Siemens
These are all the ones that expose such a functionality for backlight
effects and are present in the source (Motorola classes aren't here
yet). This is mostly used in monochrome games... those functions are
implemented with a color system that mimics some of those phones'
backlights, with Green, the most common, approximating Nokia's 3410
for example, Orange approximating Siemens' C55, etc.

Colors are fairly neutered, so "green" is actually a light green to
make the lit/unlit backlight changes easier on the eyes.