RecordStore: Add more ID checks, and save LastModified value
Fixes Wakeboard Unleashed (which also uses Motorola FunLights) and
a few others.

Since lastModified should actually be part of the rms header, but
it was not saved before, it now has to be saved at the end of the
file, after the records. Otherwise it would render older saves
incompatible.