docs: FreeJ2ME-Plus now has a Web Player!
Now this is a big commit... it's been in preparation those last
few days and the AWT frontend even received a few changes and
improvements due to this.

Most features are supported, and this also includes touch-screen
controls!

Still, this is more of a proof of concept than anything, it runs
much slower than the native binaries, and might take a while to
load depending on how fast the user's connection is, and how loaded
is CheerpJ's side... i don't intend to add much to it down the line