docs: Add a 'J2ME Documentation' section
This one is an updated version of the original archive located here:
https://github.com/nikita36078/J2ME_Docs (credits to Nikita Shakarun
for originally compiling it up).

Some JSRs were added, as well as most DoJa API specifications. It should
expand in the future.

Stylistically, the javadocs were converted to a dark mode, both to
fit into FreeJ2ME-Plus' webpage color palette, and to be easier on
the eyes.