KDDI: ezWeb: More player improvements
This commit actually simplifies some parts of KDDI's MediaPlayer
classes. No longer is a MediaManager class needed, and resources
should be tracked by the app, so all FreeJ2ME needs to concern
itself with is whether or not the players are in a valid state
when setting/unsetting resources for playback.

Some minor changes were also made to SMAFDecoder in order to support
Tetris Outdoor's packed mmf files used as a space-saving measure
(which still goes offspec after unpacking.