PlatformGraphics: Go back to per-pixel copy on copyArea() methods
Commit 174e82609d746091b285e64b7191d345fc0b6fd0 moved to a more
readable usage of System.arraycopy to copy data between areas.
However, this caused severe graphical bugs in Garfield 2: Royal
Adventure, as the game uses it to build and reuse parts of the
background tiles.

This moves us back to doing it pixel by pixel and checking bounds
for each of them. Slower, but works better.