General improvements mostly aimed at KDDI/ezPlus
RecordStore no longer returns null if there's no data in the array,
instead, it returns the empty array.

MLDDecoder now parses the CODE chunk as well, KDDI's media classes
can now handle the "dev16trp" data disposition as well, which seems
to be the official disposition that indicates that CMF audio is to
be used.

Mobile/MobilePlatform had the KDDI keys revamped, and DoJa no longer
uses the KDDI_CLR, since the CLR key appears to share the same button
as FIRE.

With these, Elevator Action for ezPlus is now perfectly playable.

PlatformGraphics also got improvements on the fillArc and fillPolygon/Triangle
methods, resulting in fixes to Landers and Bounce Tales... likely MANY
others as well.

Related to #162