Nokia: More improvements to audio playback and the OTA/OTT decoder
The OTA decoder should now support everything the most recent spec
(v3.0.0) does, aside from infinite looping patterns which can't
be supported due to MIDI limitations (SquirrelJME will be supporting
that though, so it should be a virtually perfect OTT/OTA decoder).

The source was also cleaned up and optimized thanks to needing a
nigh-complete overhaul in order to be ported into SJME, so OTA
decoding should be noticeably faster now, although it was hardly
a bottleneck to begin with.